Release V.1

+++ src/main/java/com/credibanco/assessment/card/impl/TransaccionServiceImpl.java
@@ -1,10 +1,10 @@
 package com.credibanco.assessment.card.impl;
 
-import com.credibanco.assessment.card.api.client.exception.ConflictTransactionException;
-import com.credibanco.assessment.card.api.client.exception.ConflictTransactionException2;
+import com.credibanco.assessment.card.api.client.exception.*;
 import com.credibanco.assessment.card.dto.request.AnularTransaccionRequest;
 import com.credibanco.assessment.card.dto.request.CrearTransaccionRequest;
 import com.credibanco.assessment.card.mapper.TransaccionMapper;
+import com.credibanco.assessment.card.model.Tarjeta;
 import com.credibanco.assessment.card.model.Transaccion;
 import com.credibanco.assessment.card.repository.TarjetaRepository;
 import com.credibanco.assessment.card.repository.TransaccionRepository;
@@ -18,10 +18,13 @@ import java.time.LocalTime;
 @Service
 public class TransaccionServiceImpl implements TransaccionService {
 
-  @Autowired private TransaccionRepository transaccionRepository;
-  @Autowired private TarjetaRepository tarjetaRepository;
+  @Autowired
+  private TransaccionRepository transaccionRepository;
+  @Autowired
+  private TarjetaRepository tarjetaRepository;
 
-  @Autowired private TransaccionMapper transaccionMapper;
+  @Autowired
+  private TransaccionMapper transaccionMapper;
 
   private final static String ENROLADA = "Enrolada";
 
@@ -34,10 +37,10 @@ public class TransaccionServiceImpl implements TransaccionService {
         tr.setTarjeta(tarjetaRepository.findByPan(request.getPan()).get());
         transaccionRepository.save(tr);
       }
-    }catch (ConflictTransactionException e){
-      throw new ConflictTransactionException("Error creando la transaccion "+ e.getConflictMessage());
-    }catch (ConflictTransactionException2 e){
-      throw new ConflictTransactionException2("Error creando la transaccion "+ e.getConflictMessage());
+    } catch (ConflictTransactionException e) {
+      throw new ConflictTransactionException("Error creando la transaccion " + e.getConflictMessage());
+    } catch (ConflictTransactionException2 e) {
+      throw new ConflictTransactionException2("Error creando la transaccion " + e.getConflictMessage());
     }
     return tr;
   }
@@ -45,10 +48,9 @@ public class TransaccionServiceImpl implements TransaccionService {
   @Override
   public Boolean anularTransaccion(AnularTransaccionRequest request) {
     Transaccion tr = transaccionRepository.findByNumeroReferencia(request.getNumeroRef()).orElseThrow(
-        ()-> new ConflictTransactionException("Transaccion no existe"));
-    System.out.println(tr.getTotalCompra());
-    System.out.println(request.getTotalCompra());
-    if(validarTiempo(request.getNumeroRef()) && validarMontoCompra(tr.getTotalCompra(), request.getTotalCompra())){
+        () -> new ConflictEnroledException("Transaccion con el numero de referencia " + request.getNumeroRef() + ", no existe"));
+    if (validarTiempo(request.getNumeroRef()) && validarMontoCompra(tr.getTotalCompra(), request.getTotalCompra())
+    && comprobraPan(request.getPan(), request.getNumeroRef())) {
       transaccionRepository.deleteById(tr.getId());
       return true;
     }
@@ -56,28 +58,28 @@ public class TransaccionServiceImpl implements TransaccionService {
   }
 
 
-  private Boolean validarEstadoTarjeta(String pan){
-    if(tarjetaRepository.existsByPan(pan)){
-      if(tarjetaRepository.findByPan(pan).get().getEstado().equalsIgnoreCase(ENROLADA)){
+  private Boolean validarEstadoTarjeta(String pan) {
+    if (tarjetaRepository.existsByPan(pan)) {
+      if (tarjetaRepository.findByPan(pan).get().getEstado().equalsIgnoreCase(ENROLADA)) {
         return true;
-      }else {
+      } else {
         throw new ConflictTransactionException2("Estado invalido");
       }
-    }else{
+    } else {
       throw new ConflictTransactionException("Tarjeta no existe");
     }
   }
 
-  private Boolean validarMontoCompra(Double monto1, Double monto2){
-    if(!monto1.toString().equalsIgnoreCase(monto2.toString())){
+  private Boolean validarMontoCompra(Double monto1, Double monto2) {
+    if (!monto1.toString().equalsIgnoreCase(monto2.toString())) {
       throw new ConflictTransactionException("Los Montos no son los mismos");
     }
     return true;
   }
 
-  private Boolean validarTiempo(String numeroRef){
+  private Boolean validarTiempo(String numeroRef) {
     Transaccion tr = transaccionRepository.findByNumeroReferencia(numeroRef).orElseThrow(
-        ()-> new ConflictTransactionException("Transaccion con el numero de referencia "+numeroRef+ "no existe")
+        () -> new ConflictTransactionException("Transaccion con el numero de referencia " + numeroRef + "no existe")
     );
     LocalTime time = tr.getAudit().getCreatedAt().toLocalTime();
     System.out.println(time);
@@ -85,12 +87,24 @@ public class TransaccionServiceImpl implements TransaccionService {
     System.out.println(timeNow);
     int valor = time.plusMinutes(5L).compareTo(timeNow);
     System.out.println((String.valueOf(valor)));
-    if(valor >= 0){
+    if (valor >= 0) {
       return true;
-    }else {
-      throw new ConflictTransactionException("Han pasado mas de 5 min");
+    } else {
+      throw new ConflictTransactionTimeException("Han pasado mas de 5 min, no se puede cancelar la transaccion");
     }
+
 
   }
 
+  private Boolean comprobraPan(String pan, String numeroRef) {
+    Transaccion tr = transaccionRepository.findByNumeroReferencia(numeroRef).get();
+    Tarjeta tj = tarjetaRepository.findByPan(pan).orElseThrow(()-> new ConflictTransactionException("Tarjeta no encontrada"));
+    Long idTj = tj.getId();
+    Long idTarjetaTransaccion = tr.getTarjeta().getId();
+
+    if(idTj != idTarjetaTransaccion){
+      throw new ConflictTransactionException("Los numeros de la tarjeta no coinciden");
+    }
+    return true;
+  }
 }
